requirements.txt
streamlit
requests
streamlit-geolocation
folium
streamlit-folium
pandas
altair
python-dotenv
pymysql
